Income

Data for the income information as per the SQL Murder (http://mystery.knightlab.com/walkthrough.html) and here (https://github.com/NUKnightLab/sql-mysteries). Simon Willison made an online version of the SQL game (https://sql-murder-mystery.datasette.io/sql-murder-mystery).

income

Format

A data set with two variables: ssn and annual_income.

Details

How to play the game ....

This package makes the data of the SQL Murder Mystery available in R.

There's been a Murder in SQL City! The Murder Mystery is designed to be both a self-directed lesson to learn R concepts and is also a fun game for experienced R users to solve an intriguing crime. Please read the Readme file for the details!

Examples

# Load a bit of the dataset: income
head(income)

#> # A tibble: 6 x 2
#>         ssn annual_income
#>       <int>         <int>
#> 1 100009868         52200
#> 2 100169584         64500
#> 3 100300433         74400
#> 4 100355733         35900
#> 5 100366269         73000
#> 6 100593316         16100