Interview

Data for the interviews conducted as per the SQL Murder (http://mystery.knightlab.com/walkthrough.html) and here (https://github.com/NUKnightLab/sql-mysteries). Simon Willison made an online version of the SQL game (https://sql-murder-mystery.datasette.io/sql-murder-mystery).

interview

Format

A data set with two variables: person_id and transcript.

Details

How to play the game ....

This package makes the data of the SQL Murder Mystery available in R.

There's been a Murder in SQL City! The Murder Mystery is designed to be both a self-directed lesson to learn R concepts and is also a fun game for experienced R users to solve an intriguing crime. Please read the Readme file for the details!

Examples

# Load a bit of the dataset: interview
head(interview)

#> # A tibble: 6 x 2
#>   person_id transcript                                                          
#>       <int> <chr>                                                               
#> 1     28508 "‘I deny it!’ said the March Hare.\n"                               
#> 2     63713 "\n"                                                                
#> 3     86208 "way, and the whole party swam to the shore.\n"                     
#> 4     35267 "lessons in here? Why, there’s hardly room for YOU, and no room at ~
#> 5     33856 "\n"                                                                
#> 6     82799 "\n"